SANDBOX/PY_CODE/requirements.txt
abstract-rendering
alabaster
argcomplete
astropy
Babel
backports-abc
backports.shutil-get-terminal-size
backports.ssl-match-hostname
bcolz
beautifulsoup4
binstar
bitarray
blaze
# blz
bokeh
boto
Bottleneck
butterfly
cdecimal
certifi
cffi
clyent
colorama
conda
conda-build
configobj
cryptography
cycler
Cython
cytoolz
datashape
decorator
docutils
enum34
fastcache
Flask
funcsigs
futures
gensim
gevent
gevent-websocket
greenlet
grin
h5py
idna
ipaddress
ipykernel
ipython
ipython-genutils
itsdangerous
jdcal
jedi
Jinja2
jsonschema
jupyter-client
jupyter-core
llvmlite
lxml
MarkupSafe
matplotlib
mistune
mock
multipledispatch
networkx
nltk
nose
numba
numexpr
numpy
odo
openpyxl
pandas
pandas-datareader
pathlib2
patsy
pep8
pexpect
pickleshare
Pillow
ply
prompt-toolkit
psutil
ptyprocess
py
py4j
pyasn1
pycosat
pycparser
pycrypto
pycurl
pyflakes
Pygments
pyOpenSSL
pyparsing
pytest
python-dateutil
pytz
PyYAML
pyzmq
redis
requests
requests-file
rope
ruamel-yaml
runipy
scikit-image
scikit-learn
scipy
seaborn
sense
simplegeneric
simplejson
singledispatch
six
snowballstemmer
sockjs-tornado
Sphinx
sphinx-rtd-theme
spyder
SQLAlchemy
statsmodels
sympy
tables
terminado
Theano
toolz
tornado
tornado-systemd
traitlets
ujson
unicodecsv
wcwidth
Werkzeug
wordcloud
xlrd
XlsxWriter
xlwt
